rename file to farmBot
