+++ src/components/compound/CompoundCard.tsx
@@ -3,7 +3,7 @@ import { Percent } from '@ubeswap/sdk'
 import BN from 'bn.js'
 import { useDoTransaction } from 'components/swap/routing'
 import { useToken } from 'hooks/Tokens'
-import { FarmBotSummary } from 'pages/Compound/useCompoundRegistry'
+import { FarmBotSummary } from 'pages/Compound/useFarmBotRegistry'
 import { useLPValue } from 'pages/Earn/useLPValue'
 import { Dots } from 'pages/Pool/styleds'
 import React, { useState } from 'react'
