+++ src/components/earn/ZapCard.tsx
@@ -4,7 +4,7 @@ import Loader from 'components/Loader'
 import QuestionHelper from 'components/QuestionHelper'
 import { useToken } from 'hooks/Tokens'
 import { ApprovalState } from 'hooks/useApproveCallback'
-import { FarmBotSummary } from 'pages/Compound/useCompoundRegistry'
+import { FarmBotSummary } from 'pages/Compound/useFarmBotRegistry'
 import { useLPValue } from 'pages/Earn/useLPValue'
 import { MaxButton } from 'pages/Pool/styleds'
 import React, { useCallback, useContext, useEffect, useState } from 'react'
