+++ src/pages/ProvideLiquidity/LiquidityCard.tsx
@@ -2,7 +2,7 @@ import { Token } from '@ubeswap/sdk'
 import { PoolCard } from 'components/earn/PoolCard'
 import QuestionHelper from 'components/QuestionHelper'
 import { RowBetween, RowFixed } from 'components/Row'
-import { FarmBotSummaryBase } from 'pages/Compound/useCompoundRegistry'
+import { FarmBotSummaryBase } from 'pages/Compound/useFarmBotRegistry'
 import { useLPValue } from 'pages/Earn/useLPValue'
 import React, { useState } from 'react'
 import { useTranslation } from 'react-i18next'
